+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EssenceBlockBlue.java
@@ -3,6 +3,7 @@ package com.telepathicgrunt.the_bumblezone.blocks;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import com.telepathicgrunt.the_bumblezone.blocks.blockentities.EssenceBlockEntity;
 import com.telepathicgrunt.the_bumblezone.bossbars.ServerEssenceEvent;
+import com.telepathicgrunt.the_bumblezone.configs.BzGeneralConfigs;
 import com.telepathicgrunt.the_bumblezone.items.essence.EssenceOfTheBees;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzSounds;
@@ -347,6 +348,10 @@ public class EssenceBlockBlue extends EssenceBlock {
     }
 
     public void handleGiantBubbles(ServerLevel serverLevel, BlockPos blockPos) {
+        if (!BzGeneralConfigs.blueArenaBubblesMechanic) {
+            return;
+        }
+
         long gameTime = serverLevel.getGameTime();
 
         int bubbleMoveTime = 23;