+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### **(V.7.5.13 Changes) (1.20.1 Minecraft)**
+
+#### Structures:
+Added `blueArenaBubblesMechanic` config to allow people to disable the hardcoded moving bubble blocks that spawn in that arena.
+
+
 ### **(V.7.5.12 Changes) (1.20.1 Minecraft)**
 
 #### Entities: